+++ playwright.config.ts
@@ -6,7 +6,8 @@ import { defineConfig, devices } from '@playwright/test';
  */
 // require('dotenv').config();
 
-const baseURL = `http://localhost:${process.env.npm_package_config_testPort}`;
+const port = process.env.npm_package_config_testPort || '9031';
+const baseURL = `http://localhost:${port}`;
 /**
  * See https://playwright.dev/docs/test-configuration.
  */
@@ -29,6 +30,7 @@ export default defineConfig({
 
     /* Collect trace when retrying the failed test. See https://playwright.dev/docs/trace-viewer */
     trace: 'on-first-retry',
+    permissions: ['clipboard-read']
   },
 
   /* Configure projects for major browsers */
@@ -74,10 +76,10 @@ export default defineConfig({
     //-TODO:
     // playwright tests use two servers.
     // Until an npm command is implemented to start two servers,
-    // use `npm run e2e:pw-run` to run these tests. 
+    // use `npm run e2e:pw-run` to run these tests.
   /*
     webServer: {
-    command: 'npm run e2e:start-server', 
+    command: 'npm run e2e:start-server',
     url: baseURL,
     reuseExistingServer: !process.env.CI,
   },