+++ tests/retain-additional-fields.spec.ts
@@ -2,6 +2,7 @@ import {test, expect} from '@playwright/test';
 import { MainPO } from './po/main-po';
 import path from 'path';
 import fs from 'node:fs/promises';
+import {PWUtils} from "./pw-utils";
 
 test.describe('retain-additional-fields', async () => {
   let mainPO: MainPO;
@@ -14,22 +15,7 @@ test.describe('retain-additional-fields', async () => {
   });
 
   test('should retain contained field', async ({page}) => {
-    const fileChooserPromise = page.waitForEvent('filechooser');
-    await page.getByLabel('Start from scratch').click();
-    await page.getByRole('button', {name: 'Continue'}).click();
-    await page.getByRole('button', { name: 'Import' }).click();
-    await page.getByRole('button', { name: 'Import from file...' }).click();
-    const testFile = path.join(__dirname, '../cypress/fixtures/contained-example.json');
-    const fileJson = JSON.parse(await fs.readFile(testFile, 'utf-8'));
-
-    // Start waiting for file chooser before clicking.
-    const fileChooser = await fileChooserPromise;
-    await fileChooser.setFiles(testFile);
-
-    await expect(page.locator('input#title')).toHaveValue('Contained example');
-    await page.getByRole('button', {name: 'Preview'}).click();
-    await page.getByText('View Questionnaire JSON').click();
-    const json = JSON.parse(await page.locator('mat-tab-body pre').innerText());
-    expect(json.contained).toEqual(fileJson.contained);
+    const {fileJson, fbJson} = await PWUtils.loadFile(page, '../cypress/fixtures/contained-example.json');
+    expect(fbJson.contained).toEqual(fileJson.contained);
   });
 });